minor tweaks & ActionFilterRow refactor
